requirements.txt
appnope==0.1.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
branca==0.4.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cryptography==2.9
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Fiona==1.8.13.post1
folium==0.10.1
future==0.18.2
geographiclib==1.50
geopandas==0.7.0
geopy==1.21.0
idna==2.9
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.2
json5==0.9.4
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
jupyterlab==2.1.0
jupyterlab-server==1.1.1
keyring==21.2.0
kiwisolver==1.2.0
lerc==0.1.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
ntlm-auth==1.4.0
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pprint==0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
pyparsing==2.4.7
pyproj==2.6.0
pyrsistent==0.16.0
pyshp==2.1.0
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
requests==2.23.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
retrying==1.3.3
ripe.atlas.cousteau==1.4.2
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
socketIO-client==0.7.2
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
widgetsnbextension==3.5.1
xlrd==1.2.0
